Version 0.4.2:
(1) Add Paul's new Card and exceptions classes.  The new Card class code is
more robust.  Also note that the new Card class will strip trailing blanks
for string values.

(2) move all methods from the CardList class to the Header class, except for
index_of, keys, and methods start with double underscores.  This should also
fix a bug reported by Michele when updating a binary table header.

(3) Add PCOUNT and GCOUNT as requirred cards for image extension HDU.

(4) Do byteswapping when writing out an image.

(5) Apply bscale/bzero when writing out an image.

(6) Add the ascardlist method to the Header class.

(7) Add/update NAXISi when creating an image from scratch.

(8) Modify "get" and "has_key" methods in the Header class to neglect
trailing blanks.

(9) When verify a table, neglect trailing blanks in the XTENSION value.

(10) Add the capability of referring an extension by extname (with case
insensitivity and trailing balnks neglected) and extver.  If extver is not
in the header, it will be defaulted to 1.  If there is no unique extname/extver
combination, it will throw an exception.  If only the extname is specified,
there can only be one extension with that name in the file, otherwise it
will throw an exception also.

(11) The first card of an image extension will be XTENSION = 'IMAGE   ' when
the extension is created from scratch (bug fix).

(12) Get rid of the "name" argument in BaseImageHDU.  Assign the "name"
attribute in each subclass.

(13) Change all HDU classes' argument of "cards" (a CardList instance) to
"header" (a Header instance).

(14) Change the method "__del__" to "close" in the FITS class.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@27 ed100bfc-0583-0410-97f2-c26b58777a21