This is version 0.5.0 of fits.py.  It is the first version of fits.py to be file-based, i.e. changes to the header and data will be written to disk.  The goal of this approach is to map each header and data segment of the file (or list of HDUs) to a buffer-like object, so that changes to this object will be mirrored on disk.  In version 0.5.0, this behavior only works for the header and has not yet been implemented for the data.  Therefore if changes are made to data and you wish those changes to be saved, a new FITS file will have to written.
Changes to this version of fits.py include new Buffer, File, HDUList,
and header classes (ConformingHDU, NonConformingHDU, and
CorruptedHDU); and a revised CardList class.  The functionality of the
former FITS class has now been divided between the HDUList and File
classes.  In addition a module function, called 'open', has been added
for opening a FITS file.  Descriptions of each of these classes and
their associated methods and functions can be found in the fits.html
document on the pyfits web page.

Two notable features of this version are: 1) memory-mapping is used by
default for file access and standard file I/O is used only when the
mmap module can be imported.  (Since almost all modern OSs (Unices,
Windows, Mac OS-X) provide memory-mapping, standard file I/O may be
removed from future versions.), and 2) this version expects Python
v2.1, since 'nested_scopes' is invoked and the 'warnings' framework
used to issue warning messages.

Because of these changes, a FITS file must first be opened and then
read, instead of the previous method of creating a FITS instance which
does an implicit read of the file.  For example,

>>> hdus = fits.open('good.fits').read()

creates a list of HDUs and

>>> print hdus.info()

prints a summary of the file contents.

Most other behavior has remainded the same as that of version 0.4.2.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@28 ed100bfc-0583-0410-97f2-c26b58777a21