Version 0.7.2
changes since 0.7.0.1:

(1) The tostring() call in writeHDUdata takes up extra space to store
the string object.  Use tofile() instead, to save space.

(2) Make changes from _byteswap to _byteorder, following corresponding changes
in numarray and recarray.

(3) Insert(update) EXTEND in PrimaryHDU only when header is None.

(4) Strip the trailing blanks for the comment value of a card.
(for one thing, this helps fitsdiff not to report unnecessary comment
differences".)

(5) Add seek(0) right after the __buildin__.open(0), because for the 'ab+'
mode, the pointer is at the end after open in Linux, but it is at the beginning
in Solaris.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@43 ed100bfc-0583-0410-97f2-c26b58777a21