Version 0.7.7
changes since 0.7.6:

(1) Edit _summary() in TableBaseHDU class to make the 'None' data case in
tables to work.

(2) Add 'de' to [DE] in the regex for numbers such that it is forgiving FITS
files with "slightly" illegal format in keywords of numerical values.

(3) Add the mode of "copyonwrite" for the open function.  This is equivalent
to the "readonly" mode in earlier versions.  The "readonly" mode in this
version will be a true read-only, i.e. the data can not be reassigned new
values when memmory mapping is used.  This version must be used with numarray
version 0.4.1 or later.  The default open mode is now copyonwrite.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@50 ed100bfc-0583-0410-97f2-c26b58777a21