Version 0.7.8:
(1) Modify the __getitem__ method in FITS_rec.  In order to make sure the
scaled quantities are also viewing ths same data as the original FITS_rec, all
fields need to be "touched" when __getitem__ is called.

(2) Add a new attribute mmobject for HDUList, and close the memmap object
when close HDUList object.  Earlier version does not close memmap object
and can cause memory lockup.

(3) Enable 'update' as a legitimate memmap mode.

(4) Do not print message when closing an HDUList object which is not
created from reading a FITS file.  Such message is confusing.

(5) remove the internal attribute "closed" and related method (__getattr__ in HDUList).  It is redundant.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@56 ed100bfc-0583-0410-97f2-c26b58777a21