Version 0.8.2 with the following changes:
(1) Swap back to the original byte order for image HDU data after writto().
(2) Check all table HDU's BITPIX keyword having the value 8.
(3) A new writeto() method for primary or extension HDU objects.
(4) The Column definition will take both FITS and numarray letter codes
    and check its validity in the Column definition.
(5) The arguments 'before' and 'after' in the update() method of Header
    class now accept either keyword name or index.
(6) Add the add_continue() method to Header class, so a user can add a
    CONTINUE card.

Details:

(1) When working on a little endian machine, if we create a numarray
which will have the native little endianness, and later use it to
create an HDU and write it to a file, the numarray becomes big endian
after the writeto() method.  If the user then try to use the same
array with some C-function it will have the wrong results.  This
change fixes this side effect.

(2) This checking/fixing is not done in previous version.

(3) A new writeto() method for an individual HDU (compared to the
existing one for HDUList object) simplifies a user's output interface
when only one HDU is to be written to a FITS file.  If the HDU is
primary, the FITS file will only have the primary HDU.  If the HDU
is extension, the FITS file will have a minimum primary HDU (with no data)
and the one extension HDU.

(4) The previous version of pyfits only takes FITS letter codes for its
Column's format argument.  This is inconvenient since the values in a
table HDU's columns.formats are in numarray letter codes.  This update
will allow Column to take either FITS or numarray letter codes as well
as check its validity during instanciation of the Column object instead
of checking the formats later as  implemented in previous versions.

(5) The previous version of pyfits only takes a keyword name for the
before/afetr argument in the update() method of Header class.  It now
accepts index as well.  If the index is beyond the boundary, it will
assume the boundary end point.  If before/afetr is a keyword name and
the keyword does not exist, an exception will be raised while previous
pyfits versions append the new card at the end.

(6) The add_continue() method will take a (string) value and a (string)
comment.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@67 ed100bfc-0583-0410-97f2-c26b58777a21