Version 0.9.1:
(1) When updating a Card, if the comment is specified as null string,
    it is meant so, instead of using the old one.
(2) Do not use CONTINUE if the string value can fit in one line.
    Instead, truncate the comment.  This is consistent with cases of other
    kinds of values.
(3) Fix the bug that if there is only one card with CONTINUE, it didn't work.
(4) Issue a warning message if the file size is smaller than the specified
    data size, so the user will know an input file may have been truncated.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@84 ed100bfc-0583-0410-97f2-c26b58777a21