version 0.9.8.2 (CF= convenience function)
(1) Fix the _strip() method in Header to delete the correct number of column related keywords.
     This fixes a "hung" problem when appending a table extension.

(2) Remove PCOUNT and GCOUNT from PrimaryHDU's constructor.

(3) Add the optional argument clobber to all writeto() functions and methods.

(4) Make the argument header optional in writeto(), update(), and append() CF's.  If 'header' is
     not specified and if data is numarray, it will be ImageHDU/PrimaryHDU header,
     if data is table data (i.e. FITS_rec), it will be binary table HDU header.

(5) If the HDU used in the writeto() CF is an ImageHDU, convert it to a Primary HDU.

(6) If the HDU in the append() CF is a primary HDU, convert it to a ImageHDU.

(7) Ensure that the primary HDU has EXTEND keyword when append a new HDU to a HDUList.

(8) Append new keywords after the last non-commentary card, not at the end of all keywords.

(9) If use blank cards when append/insert new keywords, account for the long (CONTINUE) card.
     (Add _use_blanks() in CardList)

(10) Improve HDUList out-of-bound index error message.

(11) Document all CF's with docstrings.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@95 ed100bfc-0583-0410-97f2-c26b58777a21