Modified test to work. They were not currently working. pyfits_test.py was failing because all the blanks padding the expected card images to 80 characters were somehow stripped. These were added back into the test file.
 table_test.py was failing because it depended on 
exact comparisons of floating point values. A couple routines to allow fuzzy 
comparisons of floating point arrays and records containing them were written.
The tests were adapted to use these helper routines (comparerecords and
compare floats).

It was noted that one of the test fits files was corrupted in the repository
(test0.fits). This will be fixed in a separate commit.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@124 ed100bfc-0583-0410-97f2-c26b58777a21