Some overhauls I started a couple weeks ago and never got back to.  I've started trying to consolidate all code that actually handles physical I/O in one place--namely the _File class, which has been renamed FITSFile.  I've also been trying to get the API just a little more consistent.
This has required moving quite a few things around.

Another note: TempHDU has been renamed RawHDU, which I think is a little more appropriate.  However, Temp/RawHDU has never lived up to its promise: the source comments claim that it exists to speed up opening files by deferring processing of HDU headers and such.  But this just leads to a bit of confusion early on.  And it never really did this deferral anyways: at the end of pyfits.open it iterated over each HDU to check its checksum (it still did the iteration even if it wasn't asked to check checksums).  Iterating over the HDUs caused them to be fully initialized anyways, so no time was saved by using the TempHDUs after all.  At best you could argue that it saved a tiny, insignificant time on I/O since it wasn't processing the HDUs much while reading from disk; but this isn't even true.  _TempHDU still had to read some headers out of the HDU text, and its method for doing this was horribly inefficient, less so even than the normal header initialization (which isn't that inefficient at all).

So all in all, nothing was gained by using TempHDU when opening files.  And even if there were a gain it would not be perceptible since the actual processing of the HDU text is a blip compared to the I/O time.  In a future changeset (if not the next one) I will remove Temp/RawHDU altogether, and also add a more flexible method for determining the type of each HDU.

git-svn-id: http://svn6.assembla.com/svn/pyfits/branches/embray_refactoring@838 ed100bfc-0583-0410-97f2-c26b58777a21