Does away with `_RawHDU` (formerly `_TempHDU`) altogether for the reasons outlined in my r838 message.  Mainly, the way it was being used didn't live up to its intent of deferring HDU initialization when opening a file.  And even more importantly, there's not much point in deferring it since it's quick anyways.  The only part that can take a long time is checksum verification, but this wasn't being deferred anyways.  And I would argue that it generally shouldn't be deferred--a user should be told if a checksum is invalid immediately after opening the file, rather that waiting until the user tries to do something with it.
If desired, we could at a later time add back in a feature to defer checksum verification only.  In the meantime, files can be opened without checksum verification (the default) and check individual HDUs' checksums with the appropriate method calls.

This fix took a while mostly because removing `_TempHDU` revealed a number of a small bugs that it was hiding.


git-svn-id: http://svn6.assembla.com/svn/pyfits/branches/embray_refactoring@839 ed100bfc-0583-0410-97f2-c26b58777a21