A change I've been wanting to make for a while.  `Header._hdutype` has gone away, and Header objects are no longer responsible for figuring out what type an HDU should be.  If it's later decided that it's still useful for Header objects to know the type of the HDU they belong to, there's no harm in adding that back in.
I much prefer the `_BaseHDU` class's `__new__()` to be responsible for
determining which HDU type to use based on the header.
Furthermore, rather than using a completely hard-coded logic, each HDU
subclass has a match function to determine whether or not it's the appropriate
type for a given header.  This system allows much more flexibility, but it's
also designed carefully enough that the matches should be unambiguous.  Even
non-standard extension types shouldn't be able to break the matching for
standard types, though there's nothing technically stopping them from
superceding the matching of standard types.



git-svn-id: http://svn6.assembla.com/svn/pyfits/branches/embray_refactoring@840 ed100bfc-0583-0410-97f2-c26b58777a21