In the process of replacing uses of the old `Header.update()` throughout the pyfits internals, I hit upon a snag: The old semantics of `Header.update()` were actually sometimes really useful: It could update an existing card or add a new one while simulataneously ensuring its position relative to other cards.  Although `Header.__setitem__` can also transparently update old/append new cards, it had no way of specifying the relative position of cards.
I thought about different ways to extend the `Header.__setitem__` semantics to
also support specifying a position update, but the options all seem cumbersome.
So it's still useful to have a method similar to the old `Header.update()`.
However, I still want to deprecate the use of `Header.update()` for this
purpose, so instead I added `Header.set()` which is similar to the old
`Header.update()`.

`Header.set()` is still useful only if you want to also ensure the position of
a keyword, but that use case comes up fairly often.  Its only difference from
the old `Header.update()` is the lack of the savecomment keyword argument.
This seems to be rarely used, and its functionality can be reproduced in other
ways if needed.


git-svn-id: http://svn6.assembla.com/svn/pyfits/branches/header-refactoring@1048 ed100bfc-0583-0410-97f2-c26b58777a21