Finishes fixing #84.  This is really sort of an edge case.  The way PyFITS works currently is that when a file is first opened, the data is not scaled and nothing is modified--this is, I guess, so that the file can be inspected as-is before the image data is scaled.
Once the image data is scaled, some of the header keywords are updated to
reflect the change.  But because the data is scaled when the file gets
rewritten to a new file, the updates to the header need to occur before the
header is written.

In the case where the data has not already been scaled, then, the _writeheader
method needs to be able to anticipate that the image data *will* get scaled,
and make the appropriate header updates ahead of time.  The data itself isn't
scaled until the data is written.

This code is a bit spaghetti-like and I might like to try to simplify it
later, but for now I was just concerned with fixing the issue.

I also replaced the previous test file with a new one that is much smaller
and that doesn't have to be gzipped, but that still exhibits the error.  With
this smaller file the test run time is closer to what it was before.


git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1136 ed100bfc-0583-0410-97f2-c26b58777a21