Refs #64.  This restores basic support for record-valued keyword cards.  Gone is the separate `RecordValuedKeywordCard`class--instead support is built right into the `Card` class.
From an architectural perspective I kind of liked how there were separate
`HierarchCard` and `RecordValuedKeywordCard` classes.  It's a nice idea to have
an interface whereby it's possible for extension classes to add support for
parsing arbitrary custom card formats.

However, in practice it never worked that well.  Hieararch cards worked pretty
well, but `RecordValuedKeywordCard` ''always'' required special support in the
`Header` class, so there was never really much of an abstraction there after
all.  If somebody actually ''asked'' for the ability to create extensions to an
abstract `Card` base class, I might try to come up with a way to make that
work.  But since I doubt anyone actually cares to have that, it's ultimately
not worth the effort.

As it is, adding support for RVKCs back into the `Card` class turned out to be
a relatively simple diff.  Though slightly kludgy, it's still far less code
than having the separate RVKC class, and ultimately a bit simpler I think.

There does seem to be a slight performance degradation with this change, which
I would like to get to the bottom of at some point.  I think it mostly just
comes from the fact that ever card needs to be checked whether or not it's a
RVKC (which was previously the case as well).  But I still suspect there are
opportunities for speedup.


git-svn-id: http://svn6.assembla.com/svn/pyfits/branches/header-refactoring@1181 ed100bfc-0583-0410-97f2-c26b58777a21