This fixes Trac #101 [writing out a scaled image fails unless the data is explicitly touched first redux].  I would consider this a "temporary" fix in that it's pretty spaghetti-code-ish.
However, after being forced to revisit, once again, the issue of writing files
I'm reminded that it still needs a lot of rework.  There's still a lot of
spaghetti code to begin with, and it's very difficult to tell just by
inspection of the code what will happen when a file gets written to,
*especially* in update mode.

The temporary fix adds an optional attribute on hdu's that flag to the
`HDUList.flush()` method that this HDU will definitely require the file to be
resized.  `_ImageBaseHDU` sets this `_resize` flag flag if the data will be
scaled.

In general, the record keeping of if an image is scaled/will be rescaled/was
rescaled needs to be improved too.  This fix adds a little record keeping,
adding an intenal `_rescaled` attribute that's set to `True` once the data has
been scaled.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1252 ed100bfc-0583-0410-97f2-c26b58777a21