Fixes an issue with mmap that was found through multidrizzle.  The use case is something like:
    >>> h = pyfits.open('filename.fits', mode='update', memmap=True)
    >>> h[1].data = h[1].data + 1
    >>> h.flush()

The problem is that `h[1].data + 1` returns a new array that is not memmap'd;
but the __array_finalize__ still coerces it to a memmap array since that's the
class of the original array.  h.flush() tries to flush the data it calls
h[1].data.flush(); but h[1].data, though using the memmap class, is no longer
backed by a mmap.mmap object, so this fails.

The solution is to view memmap arrays as a normal ndarray when returned by
`_File.readarray()`.  This way, any new arrays created from it will be based on
ndarray, *not* memmap.  This has the side effect of fixing other bugs that have
been found where certain array functions (std() for example) called on a
memmap array returned scalars wrapped in a memmap object which was confusing
and problematic for users.

This change requires a little extra work in `HDUList.flush()` to ensure that
the data array is actually backed by a memmap object.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1285 ed100bfc-0583-0410-97f2-c26b58777a21