This changeset started out trying to deal with an issue where "verification" messages were output as three warning messages: One 'warning' that was just a a header stating something like "Output verification results:", one warning at the end that was actually just a note about zero-indexing, and finally all the actual verification results lumped together in a single warning.
This was not ideal: The header and the note should not have been warnings at
all; they're purely informational.  Ideally they would be be log messages, but
for now I just had them output straight to stderr.  Then I made it so that each
validation message is output as an individual warning.

This led to the realization that all this time pyfits has been modifying the
warning system to cause warnings to be output to stdout by default instead of
stderr, explaining a great deal of strange behavior that's come up in the past.
There was probably some historical reason for that, but it's really better to
keep the expected behavior of the warnings system.

This had some implication for the tests, because while nose captures stdout it
does *not* capture stderr, leading to a mess of warning messages when running
the tests.  So some of the tests had to be updated to either catch the warnings
or to squelch stderr all together.  To this end the CaptureStdout context
manager was changed to CaptureStdio, and now captures stderr as well as stdout.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1319 ed100bfc-0583-0410-97f2-c26b58777a21