Warren was getting some strange buggy behavior regarding PyFITS' sigint handling in HDUList.flush().
After investigating the existing code for this more closely I noticed several
problems with it:

 * If an unexpected exception occurs in flush(), the default SIGINT handler
   may never be restored.
 * If a SIGINT was caught, at the end of flush() a KeyboardInterrupt is
   re-raised *before* the old SIGINT handler is restored--and thus it is never
   restoredd.
   * However this condition could never occur because the new SIGINT handler
     registered by flush() is not implemented correctly--it uses a local
     variable from the function to mark whether or not a SIGINT has been
     handled.  But that local variable is not in the handler's scope, so it
     never gets set anyways, and pyfits never re-raises the KeyboardInterrupt.

This moves all the SIGINT handling code into a new function decorator called
ignore_sigint that can be used on any function (it is not, however, reentrant--
something that might be worth fixing later).  The decorator ensures that the
old signal handler is restored at the end of the function call, and it
re-raises a KeyboardInterrupt at the end of the function if any SIGINTs were
handled.  Test included!

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1321 ed100bfc-0583-0410-97f2-c26b58777a21