A reasonable fix for #111 for now.  This removes a loop that was run over all string columns upon opening a file that ensured that strings in binary tables were right-padded with zeroes, while strings in ASCII tables were right-padded with spaces.  This was necessary when I first fixed #15 (the issue where this was first brought up.  But there have since been other fixes that ensure that ASCII tables are correctly serialized so that padding is written as spaces.
This fix ensures that when *binary* tables are written, any extra padding on
strings is written as nulls.  But it doesn't bother doing this conversion when
first opening the table.  This is because the chararray type automatically
rstrips each string anyways, so the user sees the same thing either way.  It
doesn't matter until serializing back to the file.

In any case, without the unnecessary loop when opening a table we get a
significant speedup in opening tables with string columns.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1334 ed100bfc-0583-0410-97f2-c26b58777a21