This should fix #125.
First I fixed the regression test for #119 that was incorrect, so that now it shows that merely opening and closing the file does cause its mtime to change (even if it didn't cause the contents to change).

I made a couple fixes to the Header class to improve how it reports whether or not it's been modified.  Now Header._modified is a property which also checks whether or not all the *cards* in that header have been modified directly, in addition to whether the header itself has been modified, either through insertions, deletions, or moves.

The improved modification tracking for Headers actually opened another can of worms in code that was unintentionally relying on the old buggy behavior.  I needed to re-re-re-re-re-re-re-think image scaling.  I finally went ahead and added some internal _orig_bitpix, _orig_bzero, and _orig_bscale attributes that always contain the correct values for rescaling the image.  This solved a lot of problems, and allowed me to clean up some spaghetti code.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1410 ed100bfc-0583-0410-97f2-c26b58777a21