This should solve the problem reported in #138. I'm not super happy with this, but it should work.  The `_TableLikeHDU` class is updated to have a `_data_type` class attribute, which by default is `FITS_rec`. But for `GroupHDU`s it's `GroupData`.  This allows the data for a group HDU to be created directly as a `GroupData` object, rather creating a `FITS_rec` object and wrapping it.  The old behavior is what was causing problems.
Of course, there's nothing inherently wrong with wrapping one array type in
another.  In this case the problem was specifically related to the same design
issue that's repeatedly caused reference cycle problems for pyfits tables,
which is that each `Column` object has a `.array` attribute which holds a
reference to the recarray that contains the column.  This creates all kinds of
reference cycle issues.

Eventually PyFITS will switch over to using the Astropy table interface, which
should free up a lot of these issues.  But in the meantime this will have to
do.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1454 ed100bfc-0583-0410-97f2-c26b58777a21