Starting to do a little cleanup--I'm still not sure if scripts/fitsdiff will be the final living place for fitsdiff's documentation--depends on how easily I can get it into the pyfits docs.
Some progress on cleaning up fitsdiff--not working yet though.

More progress on updating the fitsdiff options--still broken though.

Mostly done the option handlign and help formatting.  Now to actually do stuff with the options...

Finishes cleaning up the fitsdiff script to a point I'm happy with for now.  Currently doesn't pass any arguments to the fitsdiff function other than the filenames, but I'll work on that.  Also, I kind of wanted to change its handling of directory comparisons to work more like the normal diff command, in that the filenames in the directories would also be compared.  But for now I decided to keep with the existing functionality of fitsdiff.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1529 ed100bfc-0583-0410-97f2-c26b58777a21