Fixed #146 in the manner specified in the ticket, by adding a new global variable.
Writing a test for this gave me an opportunity to slightly rethink how
whitespace is currently handled, and streamline things a bit so that it's
handled more consistently.  Now, if you manually assign a value to a header
containing trailing whitespace, that whitespace is treated as intentional and
is *preserved* when writing the file::

    >>> h['FOO'] = 'Bar        '
    >>> print h.cards['FOO']
    FOO     = 'Bar        '

However, when accessing the value through the header the whitespace is still
stripped like normal::

    >>> h['FOO']
    'Bar'

This is the same as though the value had been read directly from a file::

    >>> c = pyfits.Card.fromstring("FOO     = 'Bar        '")
    >>> c.value
    'Bar        '
    >>> h = pyfits.Header()
    >>> h.append(c)
    >>> h['FOO']
    'Bar'

This is unless of course ``pyfits.STRIP_HEADER_WHITESPACE = False``.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1566 ed100bfc-0583-0410-97f2-c26b58777a21