Add a regression test for the forthcoming HDUList.fromstring method
Finished implementing HDUList.fromstring.  Expanded the test for it to test opening and reading files of several different types, which helped locate a few additional bugs.  Still not sure if this is perfect, but it seems good enough for now. It should be emphasized that in addition to strings, this should work with any object that implements the buffer interface, including ndarrays.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1594 ed100bfc-0583-0410-97f2-c26b58777a21