Some changes that I think will be necessary for this to work on Python 3, to handle strings incoming as bytes objects.
This test was failing in Python 3, I think due to a Numpy bug related to dtypes but I'm not sure--doing a column by column comparison for tables gets around it.

Converting the header to a string here was causing problems on Python 3, since str() no longer returns an object that implements the buffer interface.  Still need to think about smoothing that out a bit.  But in the meantime, the str() call was unnecessary since _BaseHDU.fromstring can take a Header object already.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1595 ed100bfc-0583-0410-97f2-c26b58777a21