Fixes some pretty bad inefficiencies that occurred when writing compressed data, and in general tables containing variable-length arrays where the heap data is in an in-memory buffer.
First of all, after compressing the data we truncate the compressed data array
which usually allocates enough memory for the worst-case compression scenario,
but once compression is done it can usually free most of that memory.

Second, when reading VLA data we were slicing a Python buffer object for every
single row of the VLA column. But slicing a buffer, contrary to my assumption,
doesn't return a new buffer but instead copies the buffer data into a new
string.  Doing this thousands of times (once for each compressed tile) was
enormously inefficient.  Now we support holding the heap data in a Numpy
array that can be sliced many times at relatively low cost.

However, there are still some inefficiencies here, particularly in
the BinTableHDU._binary_table_byte_swap method.  I need to see if I can fix it
up a bit as well.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@1923 ed100bfc-0583-0410-97f2-c26b58777a21