Moved truncation of over-allocated table+heap arrays to the C code.
I'm not sure why I ever expected the existing code to work as written:
creating an array using PyArray_SimpleNewFromData never gives the array
ownership of the data, which mean that writing compressed data always
required making a copy with `np.resize()` it seems like.

Another implication of this is that the created Numpy array doesn't
own the data (as in, the memory for the data was not allocated by
Numpy itself), then when the Numpy array is dereferenced it never
frees the data.  So I think there might be a memory leak when writing
compressed data, but I need to test this.  It's possible such a memory
leak has already been reported but I still need to check on that too.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2257 ed100bfc-0583-0410-97f2-c26b58777a21