Adds support for writing compressed images using the Q format for compressed data columns where appropriate.  As the comments in the code say, the heuristic used to determined whether the Q format should be used is imperfect, but it is the same heuristic used by CFITSIO so at least it's consistent.
This also fixes a long-standing bug where it was assumed that each column in a compressed image BINTABLE is 8 bytes wide.  In fairness, that was previously a reasonable assumption (most of the columns were either type P or D), but the standard does permit other columns in the table besides the ones explicitly defined by the standard.  And now of course with support for the Q type columns can be 16 bytes wide as well.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2260 ed100bfc-0583-0410-97f2-c26b58777a21