Changed the behavior of the internal HDU property `_data_loaded` so that it returns True even if `self.data is None`.  Sometimes it's useful to check if the HDU's data has been read/processed/replaced even if its currently value is now `None`.  Added a new property called `_has_data` with the same behavior as the old `_data_loaded`.
This had some surprising side-effects: in the `._summary` method of some
of the HDU classes there was a check that previously could never passed
(because it required `self._data_loaded and self.data is None` which was
previously impossible) that returned an array's format string as an
empty string if contains zero data.  Making this change caused that test
to start passing and changed the output of the `pyfits.info()` function.
I thought this new behavior seemed reasonable so I decided to change the
tests and not keep the old behavior.  But I needed a few additional
tweaks (especially in GroupsHDU) to make sure the new behavior applied
consistently.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2272 ed100bfc-0583-0410-97f2-c26b58777a21