This change resolves an issue where table data was unnecessary read from a file and/or put through a time-consuming processes of checking to see if each column needed to byteswapped, even if the table data was completely unmodified.  Furthermore, the byteswapping code had a bug where it could cause the heap data for VLAs to be rewritten, but in the wrong order.  This bug is difficult to write a test for within pyfits because the way pyfits writes VLA columns it never interleaves the heap data for two columns.  I have a sample file that is written in such a manner, however, and this change was able to fix issues with writing compressed data from that file to a new file.  Still, before I can call this issue 100% fixed I will try to contrive a test that exhibits the problem.
This added a simple method for copying data direct from one file to
another, and without reading the source bytes entirely into memory first
(which is the old behavior--probably required an inordinate amount of
memory to be used when writing files in certain circumstances).

This also required finally adding a `._data_needs_rescaling` attribute
on ImageHDUs to clarify up front, even before an image's data has been
read/rescaled, that upon writing to a new file the data *will* need to
be rescaled (see track issue #101 and friends).  Later a similar
attribute will need to be added to table HDUs, but only once the PR for
scaling table columns has been merged (there is one such fix in
progress).

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2273 ed100bfc-0583-0410-97f2-c26b58777a21