As a prerequisite to improving the consistency of how scaled table columns are handled I finally got around to refactoring the FITS_rec.field method a bit.  This is work I had been avoiding in preference to just working on the existing plan of redoing table support on top of Astropy tables.
That said, this was still relatively easy to do, and I think that the
new, cleaner code will be significantly easier to transfer over to a new
table implementation. It will also lend itself better to supporting
backwards compatibility with the recarray-based approach which will be
necessary for the first few releases that include any new table
implementation.

One major difference to the internal data structures introduced by this
fix is that *all* of the table columns are added to the
FITS_rec._convert list of converted columns, even if they don't require
any scaling or data type conversion.  In the non-conversion case the
array stored in ._convert is just the direct slice of the recarray
returned by recarray.field.  I'm still a little concerned whether this
change leads to memory leaks and that's something I need to test for
before merging this new code.  If it does cause memory leaks it would be
due to a reference cycle, and I think I can find a workaround.

There are a few other TODO items in this commit that need to be
addressed before merging as well.

This commit contains a fix for a bug that hadn't revealed itself until
making these other changes (but was necessary to fix for all the tests
to pass).  This is specifically the fact that the `FITS_rec.copy()`
method (inherited from `ndarray.copy()`) did not perform a deep copy of
all of a FITS_rec's attributes, including the arrays in the `._convert`
list.  This meant that when copying a `FITS_rec` with the copy method
the new array would have its own copy of the *raw* data, but the
converted arrays would remain shared.  This is especially bad in the new
case where slices of *all* of the columns are stored in `._convert`.
This meant that any update to the new table updated the old table as
well (likewise they shared the same column definitions).  So now
`FITS_rec.copy()` is overridden to ensure that all of the table's
attributes get a deep copy.  This in turn revealed that since
`Column.__new__` expects a minimum number of arguments,
`copy.(deep)copy()` did not work on it without first defining custom
`__(deep)copy__` methods.

*whew!*

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2319 ed100bfc-0583-0410-97f2-c26b58777a21