Moved all the real functionality of new_table into a new FITS_rec.from_columns class method.  This makes vastly more sense, in that most of what new_table did was to create a new FITS_rec object and fill its columns based on the format specification in a list of Column objects.  It makes sense that such a task belong to the FITS_rec class itself--some of the code for doing this overlaps with other code already in FITS_rec, especially for converting from the Numpy-native data formats to the FITS-native data format.  Not all of the overlapping code has been refactored, as some of it will take some thinking, but there are notes for where it seems applicable.
The rest of new_table's responsibilities are to instantiate a
BinTableHDU or TableHDU depending on what was passed into its `tbtype`
argument.  Now, new_table just creates the data array with
FITS_rec.from_columns and uses the normal constuctors for those classes
to instantiate the HDU.  The code is now vastly cleaner and simpler to
follow.  I also modified new_table to allow passing in the `BinTableHDU`
and `TableHDU` classes themselves to specify what type of table to
create, rather than the string.  I think this approach is better
overall, but that's subjective.  I also got rid of the `eval()` call
which I always felt was pretty ugly.

In the process of moving the code around I noticed one slight defect of
the old table creation code: In binary tables, the default values in
string columns were always set to the string '0.0' which seemed odd and
unintuitive.  This has been fixed so that string values are '' (empty
strings) by default.

Finally, there's still some work to be done here--this change is not
quite complete.  For one: The old new_table code had some logic related
to scaled columns.  For now I've removed that logic since it did not
seem to affect the results of any existing tests.  That doesn't mean it
didn't have a purpose though--I need to review exactly what it was meant
for, determine if it is *still* needed, and include some tests that
exercise it.  There is also the matter of handling TDISPn and TNULLn
when converting between ASCII columns and binary columns.  These fields
have different semantics depending on the table type, so the column
conversions need a way to account for those differences.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2329 ed100bfc-0583-0410-97f2-c26b58777a21