Manual merge of https://github.com/astropy/astropy/pull/906 with a few small changes, and a few larger changes (read: hacks) to get this change integrated with my new half-complete table refactoring.  The result is a big huge mess--even worse than I expected--illustrating why I wanted to get further with the refactoring in the first place before incorporating this change.  But c'est la vie...
Another substative change worth pointing is out is I added an internal
switch so that table columns are read as uints *only* if the file was
opened with uint=True, as is the case for images.  How this is handled
will be revisited in a future version (really it should be on by
default, as even the FITS Standard discusses this convention in newer
versions).

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2339 ed100bfc-0583-0410-97f2-c26b58777a21