Fixed two bugs:
 1. Writing a binary table HDU with compression left the original HDU's
    data byteswapped (and hence mildly "corrupted").

 2. Compressed image HDUs did not correctly compute the ZHECKSUM and
    ZDATASUM header keywords for the checksums of the original image.

This involved a long-needed breaking up of
BinTableHDU._binary_table_byte_swap into the actual byte swapping
operations, and the parts that wrote to a file.  Now that method just
does byte-swapping, and the details of writing the raw data to a file
object (including the heap) are handled in
BinTableHDU._writedata_internal where it belongs.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2435 ed100bfc-0583-0410-97f2-c26b58777a21